# Optimize LodePNG