Refactor and optimize existing LodePNG functions

These changes decrease the size of the compiled LodePNG code and speed up optimizing PNGs using the --allfilters switch

On MSYS2 MINGW64 gcc 12.2.0, LodePNG code is

-  10.62% smaller (170386 bytes -> 152288 bytes, 17.67kB smaller)
- ~11.17% faster, using the command ect -4096 -strip --allfilters "[neat image.png](https://user-images.githubusercontent.com/97068837/189489992-cc524dd2-8cce-4277-9581-3d24852e009c.png)" (47.2513s -> 41.9741s)

On MSYS2 CLANG64 clang 15.0.0, LodePNG code is

- 6.67% smaller (169356 bytes -> 158052 bytes, 11.04kB smaller)
- ~11.83% faster, using the command ect -4096 -strip --allfilters "[neat image.png](https://user-images.githubusercontent.com/97068837/189489992-cc524dd2-8cce-4277-9581-3d24852e009c.png)" (42.8687s -> 37.7980s)

There's a ~0.5% speed increase for the --allfilters-b switch, otherwise there is no measurable speed difference optimizing PNGs not using --allfilters or --allfilters-b